Cache wird nun nach jeder Sitzung gelöscht, eigene Nachrichten werden rechts angezeigt

+++ app/src/main/java/de/hs_niederrhein/chat/hsnrchat/ClientServerCommunicator.java
@@ -80,7 +80,8 @@ public class ClientServerCommunicator extends ServerCommunicator {
 
     @Override
     public void onNewMessage(long userId, short roomId, String message) {
-        db.insertMessage((int)roomId, message, (int)userId);
+        db.insertMessage((int)roomId, message, userId);
+
 
     }
